network: move LruHashSet to network crate utils
